refactor review comments
