+++ packages/docusaurus-utils/src/__tests__/jsUtils.test.ts
@@ -7,22 +7,7 @@
 
 import {jest} from '@jest/globals';
 import _ from 'lodash';
-import {
-  removePrefix,
-  mapAsyncSequential,
-  findAsyncSequential,
-} from '../jsUtils';
-
-describe('removePrefix', () => {
-  it("is no-op when prefix doesn't exist", () => {
-    expect(removePrefix('abcdef', 'ijk')).toBe('abcdef');
-    expect(removePrefix('abcdef', 'def')).toBe('abcdef');
-    expect(removePrefix('abcdef', '')).toBe('abcdef');
-  });
-  it('removes prefix', () => {
-    expect(removePrefix('prefix', 'pre')).toBe('fix');
-  });
-});
+import {mapAsyncSequential, findAsyncSequential} from '../jsUtils';
 
 describe('mapAsyncSequential', () => {
   function sleep(timeout: number): Promise<void> {
